+++ src/AnomalyDetectors/Loda.php
@@ -15,7 +15,7 @@ use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Other\Helpers\Stats;
 use Rubix\ML\Other\Helpers\Params;
 use Rubix\ML\Other\Helpers\Verifier;
-use Rubix\ML\Other\Traits\RanksSingle;
+use Rubix\ML\Other\Traits\ScoresSingle;
 use Rubix\ML\Other\Traits\PredictsSingle;
 use Rubix\ML\Specifications\DatasetIsNotEmpty;
 use Rubix\ML\Specifications\DatasetHasDimensionality;
@@ -44,7 +44,7 @@ use const Rubix\ML\LOG_EPSILON;
  */
 class Loda implements Estimator, Learner, Online, Ranking, Persistable, Stringable
 {
-    use PredictsSingle, RanksSingle;
+    use PredictsSingle, ScoresSingle;
 
     /**
      * The minimum number of histogram bins.
@@ -321,7 +321,7 @@ class Loda implements Estimator, Learner, Online, Ranking, Persistable, Stringab
      */
     public function predict(Dataset $dataset) : array
     {
-        return array_map([$this, 'decide'], $this->rank($dataset));
+        return array_map([$this, 'decide'], $this->score($dataset));
     }
 
     /**
@@ -331,7 +331,7 @@ class Loda implements Estimator, Learner, Online, Ranking, Persistable, Stringab
      * @throws \RuntimeException
      * @return float[]
      */
-    public function rank(Dataset $dataset) : array
+    public function score(Dataset $dataset) : array
     {
         if (!$this->r or !$this->histograms or !$this->threshold) {
             throw new RuntimeException('Estimator has not been trained.');
@@ -344,6 +344,21 @@ class Loda implements Estimator, Learner, Online, Ranking, Persistable, Stringab
             ->transpose();
 
         return $this->densities($projections);
+    }
+
+    /**
+     * Return the anomaly scores assigned to the samples in a dataset.
+     *
+     * @deprecated
+     *
+     * @param \Rubix\ML\Datasets\Dataset $dataset
+     * @return float[]
+     */
+    public function rank(Dataset $dataset) : array
+    {
+        trigger_error('Deprecated, use score() instead.', E_USER_DEPRECATED);
+
+        return $this->score($dataset);
     }
 
     /**