+++ src/NeuralNet/Layers/Dense.php
@@ -21,7 +21,7 @@ use Generator;
  * Dense (or *fully connected*) hidden layers are layers of neurons that connect to each node
  * in the previous layer by a parameterized synapse. They perform a linear transformation on
  * their input and are usually followed by an Activation layer. The majority of the trainable
- * parameters in a standard feed forward neural network are contained within Dense hidden layers.
+ * parameters in a standard feed-forward neural network are contained within Dense hidden layers.
  *
  * @category    Machine Learning
  * @package     Rubix/ML